+++ src/main/java/roomescape/dao/ReservationDao.java
@@ -1,0 +1,11 @@
+package roomescape.dao;
+
+import java.util.List;
+import roomescape.Reservation;
+
+public interface ReservationDao {
+
+    long save(final Reservation reservation);
+    List<Reservation> findAll();
+    void delete(long id);
+}
