+++ src/main/java/roomescape/repository/ReservationRepository.java
@@ -5,7 +5,7 @@ import java.util.List;
 import roomescape.Reservation;
 
 public interface ReservationRepository {
-    long save(final Reservation reservation);
-    List<Reservation> findAll();
-    void delete(final long id);
+    long add(final Reservation reservation);
+    List<Reservation> getAll();
+    void remove(final long id);
 }
